+++ CHANGELOG.md
@@ -7,6 +7,8 @@
    available starting with revision 2 of `VK_EXT_debug_report`.
  * All `*_EXTENSION_SPEC_VERSION` have been renamed to `*_SPEC_VERSION`.
  * Add `VK_VALIDATION_CHECK_SHADERS_EXT` to `VkValidationCheckEXT`.
+ * Support revision 8 of `VK_EXT_debug_report`. Revision 7 is not explicitly supported, because
+   there is no official `vulkan.h` with just that revision.
 
 ## Version 0.14.1, released on 21.05.2017